Refactor: cleanup, init defaults, and reduce includes
Improved member initialization with in-class defaults, removed redundant includes to reduce dependencies, updated forward declarations, and refactored code for clarity and maintainability. Commented out heavy test scene generation and improved OpenGL resource handling. Minor syntax and type consistency fixes throughout.

+++ opengl/src/physics/broadphase/broadphase_manager.cpp
@@ -1,15 +1,15 @@
 #include "pch.h"
 #include "broadphase_manager.h"
-#include "game_object.h"
 #include "rigid_body.h"
 #include "tri.h"
+#include "bvh/treetree_query.h"
 
 void BroadphaseManager::init(PhysicsWorld* world, RuntimeCaches* caches, std::vector<Tri>* terrainTris) {
     this->caches = caches;
     this->terrainTriangles = terrainTris;
 
     SlotMap<RigidBody, RigidBodyHandle>* bMap = caches->bodies.sm; // sm=slotmap
-    uint32_t slotCap = bMap->dense().capacity();
+    size_t slotCap = bMap->dense().capacity();
 
     awakeBvh.init(world, caches, slotCap);
     asleepBvh.init(world, caches, slotCap);
